feat(wit): refactor wit a bit
