+++ rust/src/lib.rs
@@ -1,4 +1,5 @@
 mod tracing;
+mod types;
 
 pub use tracing::*;
