+++ wit/world.wit
@@ -1,6 +1,7 @@
 package wasi:otel@0.2.0-draft;
 
 world imports {
+    import types;
     import tracing;
 }
